+++ lib/main.dart
@@ -58,63 +58,3 @@ class DetailPage extends StatelessWidget {
     );
   }
 }
-
-// 수홍's 자기소개 페이지
-class Suhong extends StatelessWidget {
-  Suhong({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      body: Center(
-        child: Text("Suhong's page"),
-      ),
-    );
-  }
-}
-
-// 우준's 자기소개 페이지
-class Woojun extends StatelessWidget {
-  Woojun({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      body: Center(
-        child: Text("Woojun's page"),
-      ),
-    );
-  }
-}
-
-// 현우's 자기소개 페이지
-class Hyunwoo extends StatelessWidget {
-  Hyunwoo({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      body: Center(
-        child: Text("Hyunwoo's page"),
-      ),
-    );
-  }
-}
-
-// 성학's 자기소개 페이지
-class Sunghak extends StatelessWidget {
-  Sunghak({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      body: Center(
-        child: Text("Sunghak's page"),
-      ),
-    );
-  }
-}
