+++ lib/suhong.dart
@@ -1,0 +1,17 @@
+// 수홍's 자기소개 페이지
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Suhong extends StatelessWidget {
+  Suhong({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Center(
+        child: Text("Suhong's page"),
+      ),
+    );
+  }
+}
