+++ lib/sunghak.dart
@@ -1,0 +1,17 @@
+// 성학's 자기소개 페이지
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Sunghak extends StatelessWidget {
+  Sunghak({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Center(
+        child: Text("Sunghak's page"),
+      ),
+    );
+  }
+}
