+++ lib/woojun.dart
@@ -1,0 +1,17 @@
+// 우준's 자기소개 페이지
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Woojun extends StatelessWidget {
+  Woojun({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Center(
+        child: Text("Woojun's page"),
+      ),
+    );
+  }
+}
